feat(#195): adopt koin annotation for dependency injection (#120)
