+++ app/src/main/java/com/brainwallet/navigation/LegacyNavigation.kt
@@ -10,8 +10,7 @@ import androidx.core.net.toUri
 import com.brainwallet.BuildConfig
 import com.brainwallet.R
 import com.brainwallet.data.repository.LtcRepository
-import com.brainwallet.data.source.RemoteApiSource
-import com.brainwallet.di.getKoinInstance
+import com.brainwallet.di.AppModule.getKoinInstance
 import com.brainwallet.presenter.activities.BreadActivity
 import com.brainwallet.ui.BrainwalletActivity
 import kotlinx.coroutines.CoroutineScope
