+++ app/src/main/java/com/brainwallet/ui/screens/home/receive/ReceiveDialogViewModel.kt
@@ -21,8 +21,10 @@ import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.flow.updateAndGet
 import kotlinx.coroutines.launch
+import org.koin.android.annotation.KoinViewModel
 
 //todo: wip
+@KoinViewModel
 class ReceiveDialogViewModel(
     private val settingRepository: SettingRepository,
     private val ltcRepository: LtcRepository,
