+++ app/src/main/java/com/brainwallet/ui/screens/inputwords/InputWordsViewModel.kt
@@ -16,7 +16,9 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import org.koin.android.annotation.KoinViewModel
 
+@KoinViewModel
 class InputWordsViewModel : BrainwalletViewModel<InputWordsEvent>() {
 
     private val _state = MutableStateFlow(InputWordsState())
