+++ app/src/main/java/com/brainwallet/ui/screens/ready/ReadyViewModel.kt
@@ -1,7 +1,9 @@
 package com.brainwallet.ui.screens.ready
 
 import com.brainwallet.ui.BrainwalletViewModel
+import org.koin.android.annotation.KoinViewModel
 
+@KoinViewModel
 class ReadyViewModel : BrainwalletViewModel<ReadyEvent>() {
 
     override fun onEvent(event: ReadyEvent) {
