+++ app/src/main/java/com/brainwallet/ui/screens/setpasscode/SetPasscodeViewModel.kt
@@ -11,7 +11,9 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.flow.updateAndGet
 import kotlinx.coroutines.launch
+import org.koin.android.annotation.KoinViewModel
 
+@KoinViewModel
 class SetPasscodeViewModel : BrainwalletViewModel<SetPasscodeEvent>() {
 
     private val _state = MutableStateFlow(SetPasscodeState())
